safari-600.7-branch/LayoutTests/http/tests/security/contentSecurityPolicy/source-list-parsing-no-semicolon.html

<!DOCTYPE html>
<html>
<head>
<script src='resources/multiple-iframe-test.js'></script>
<script>
var tests = [
    ['yes', 'default-src \'self\' script-src example.com', 'resources/script.js'],
    ['yes', 'script-src \'self\' object-src \'self\' style-src *', 'resources/script.js'],
];
</script>
</head>
<body onload="test()">
    <p>If a web author forgets a semicolon, we should do our best to warn them
    that the policy they've defined is probably not what they intended.</p>
